Rename vanilla.js -> transformClass.js
