+++ packages/babel-plugin-transform-classes/src/transformClass.js
